--- oauth2-test/spring-security-oauth2-resource-server-test.gradle
@@ -1,16 +1,0 @@
-apply plugin: 'io.spring.convention.spring-module'
-
-dependencies {
-	compile project(':spring-security-test')
-	compile project(':spring-security-oauth2-resource-server')
-	compile project(':spring-security-oauth2-jose')
-	compile project(':spring-security-oauth2-client')
-	
-	compile 'org.springframework:spring-test'
-	
-	testCompile project(':spring-security-config')
-	
-	optional 'org.springframework:spring-webflux'
-	
-	provided 'javax.servlet:javax.servlet-api'
-}
