--- oauth2-test/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.test.context.TestExecutionListener = \
-	org.springframework.security.test.context.support.WithSecurityContextTestExecutionListener,\
-	org.springframework.security.test.context.support.ReactorContextTestExecutionListener
