--- oauth2-test/template.mf
@@ -1,19 +1,0 @@
-Implementation-Title: org.springframework.security.oauth2.test
-Implementation-Version: ${version}
-Bundle-SymbolicName: org.springframework.security.oauth2.test
-Bundle-Name: Spring Security OAuth2 Test
-Bundle-Vendor: SpringSource
-Bundle-Version: ${version}
-Bundle-ManifestVersion: 2
-Ignored-Existing-Headers:
- Import-Package,
- Export-Package
-Import-Template:
- org.apache.commons.logging.*;version="${cloggingRange}",
- org.springframework.security.core.*;version="${secRange}",
- org.springframework.security.authentication.*;version="${secRange}",
- org.springframework.security.oauth2.server.resource.*;version="${secRange}",
- org.springframework.security.web.*;version="${secRange}",
- org.springframework.beans.factory;version="${springRange}",
- org.springframework.util;version="${springRange}",
- javax.servlet.*;version="0"
